pyproject.toml
[project]
name = "gov-dossier-toelatingen"
version = "0.1.0"
description = "PROV-based dossier management API for toelatingen beschermd erfgoed"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    # For PostgreSQL (uncomment when ready):
    # "asyncpg>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.26.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
postgres = [
    "asyncpg>=0.29.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["gov_dossier_engine*", "gov_dossier_toelatingen*", "gov_dossier_app*"]
